Split playback stuff from mcvqoe.gui
This moves the check_audio function into mcvqoe/hardware/PTT_play and
calls it single_play. This function can now play audio using the given
objects and, optionaly, play back the result. also added the test.wav
clip to the package and read in the data correctly so it should always
be found.
This fixes issue #4