Fix problem in datafilenames with python tests
When the code was transitioned to python, the log format was changed so
that instead of each log for a measurement starting with:
>> Test started at [...]
It now has the test name instead of "Test" this breaks datafilenames as
it checks this to determine how to look for files.
This fix fixes this by also treating the Python test names the same way
that "Test" was prevously treated.
Note : I have not tested this as log parsing also seems to be broken :(