Update sync with bidirectional option
Updates sync to have bidirectional be an option and default it to false
in all cases. This is diffrent than the prevous default behavior where
sync with no arguments would be a bidirectional sync. This wasn't a big
deal when all our data came from the lab, but now, with simulations,
this is slowing things down.