Recovery-added bisect variables, labelled obvious bugs
Recovery has not been tested in some time. I ran a test and it failed.
When I looked into it there are a lot of major issues here. I do not see
it as a priority right now compared to making access time tests shorter
with bisect_midpoint, but these need to be addressed at some point